refactor: calculateDiscountAmount util 함수로 분리 및 discountAmount 전역상태 삭제
