+++ src/hooks/useAdditionalShippingFeeArea.ts
@@ -1,20 +1,15 @@
 import {
-  ADDITIONAL_SHIPPING_FEE,
-  COUPON_DISCOUNT_TYPE,
-} from "@/constants/system";
-import {
   additionalShippingFeeAreaState,
-  discountAmountState,
   selectedCouponsState,
 } from "@/store/atoms/atoms";
 import { useRecoilState, useRecoilValue } from "recoil";
+
+import { COUPON_DISCOUNT_TYPE } from "@/constants/system";
 
 const useAdditionalShippingFeeArea = () => {
   const [isSelected, setSelected] = useRecoilState(
     additionalShippingFeeAreaState
   );
-  const [discountAmount, setDiscountAmount] =
-    useRecoilState(discountAmountState);
 
   const selectedCoupons = useRecoilValue(selectedCouponsState);
 
@@ -28,13 +23,6 @@ const useAdditionalShippingFeeArea = () => {
     ) {
       return;
     }
-
-    if (isSelected) {
-      setDiscountAmount(discountAmount + ADDITIONAL_SHIPPING_FEE);
-      return;
-    }
-
-    setDiscountAmount(discountAmount - ADDITIONAL_SHIPPING_FEE);
   };
 
   return { isSelected, handleSelect };
